requirements.txt
fastmcp
psycopg2-binary
python-dotenv
rich
google-generativeai
